Tokenizers fall back to BPE if unregistered
Now that the library has been tested more extensively it's not so
valuable to fail on tokenizers we haven't encountered before.

We still need to check what happens with tokenizers that use a different
model, or those that haven't been ported from their original
implementations.

An alternative would be to expose a registration mechanism, as discussed
in #63.